invoked_by_apigw_cloudformation returns 200 True
